Se ajusto para que se pueda guardar Gastos e Ingresos

+++ app/src/main/java/software/mys/guardaditoapp/data/repositories/TransactionRepository.kt
@@ -11,15 +11,19 @@ class TransactionRepository(
     private val accountDao: AccountDao
 ) {
     fun insertTransaction(transaction: TransactionEntity) {
+        //LOGINCA DE AGREGAR UNA TRANSACCION
+        //La transaccion se realizo y se debe actualizar el account balance.
         transactionDao.insertMovement(transaction)
         val account = accountDao.getAccountById(transaction.accountId)
         when (transaction.type) {
             TransactionTypeEntity.EXPENSE -> {
-                account.balance -= transaction.amount
+                account.balance -= transaction.amount.toDouble()
             }
+
             TransactionTypeEntity.INCOME -> {
-                account.balance += transaction.amount
+                account.balance += transaction.amount.toDouble()
             }
+
             TransactionTypeEntity.TRANSFER -> {
 
             }
